requirements.txt
streamlit>=1.36.0
databricks-sql-connector>=3.0.0
databricks-sdk>=0.30.0
